+++ plugins/html/CHANGELOG.md
@@ -1,3 +1,7 @@
+v2.1.4
+- Reemitting src/href/styles assets
+- Added minimum-scale=1.0 to viewport
+
 v2.1.3
 - Updated npm packages